fixing combat

+++ Game/Creature.pde
@@ -1,6 +1,6 @@
 public class Creature {
   private Dungeon dungeon;
-  private float health, energy;
+  private double health, energy;
   private int x, y; //position in the floor map
   private int level;
   private boolean enemy;
@@ -8,7 +8,7 @@ public class Creature {
   long nextMovement;
   PImage enemyImage, heroImage; 
 
-  public Creature(float health, float energy, int x, int y, Dungeon dungeon, boolean enemy, Random rng) {
+  public Creature(double health, double energy, int x, int y, Dungeon dungeon, boolean enemy, Random rng) {
     this.x = x;
     this.y = y;
     this.health = health;
@@ -25,19 +25,19 @@ public class Creature {
     return dist(x, y, other.x, other.y);
   }
 
-  public float getHealth() {
+  public double getHealth() {
     return health;
   }
 
-  public void setHealth(float health) {
+  public void setHealth(double health) {
     this.health = health;
   }
 
-  public float getEnergy() {
+  public double getEnergy() {
     return energy;
   }
 
-  public void setEnergy(float energy) {
+  public void setEnergy(double energy) {
     this.energy = energy;
   }
