+++ Game/Hero.pde
@@ -2,7 +2,7 @@ public class Hero extends Creature {
   long nextMovement;
   int maxHealth = 20;
   boolean attemptAttacking;
-  public Hero(float health, float energy, int x, int y, Dungeon dungeon, boolean enemy, Random rng) {
+  public Hero(double health, double energy, int x, int y, Dungeon dungeon, boolean enemy, Random rng) {
     super(health, energy, x, y, dungeon, enemy, rng);
     nextMovement = millis() + 120;
     attemptAttacking = false;
